Initial creation of repository
This repository contains a python file used to track the status of the HSC
LSST code merge. The python file contains a dictionary which maps a HSC
tickt number to a relivant message. The python file is then used to
generate a webpage with relivant info on the status of the HSC stack